Move file loader to it's own space and updated readme
