+++ examples/allFeatures/package.json
@@ -9,7 +9,7 @@
     "ini": "^3.0.1",
     "json5": "^2.2.1",
     "toml": "^3.0.0",
-    "typescript": "^4.9.3",
+    "typescript": "^4.9.4",
     "x2js": "^3.4.4",
     "yaml": "^2.1.3"
   },
@@ -17,7 +17,7 @@
     "@types/cson": "^7.20.1",
     "@types/hjson": "^2.4.3",
     "@types/ini": "^1.3.31",
-    "@types/node": "^18.11.11",
+    "@types/node": "^18.11.12",
     "ts-node": "^10.9.1"
   }
 }
