+++ examples/webserver/package.json
@@ -5,11 +5,11 @@
   "dependencies": {
     "axios": "^1.2.1",
     "express": "^4.18.2",
-    "typescript": "^4.9.3"
+    "typescript": "^4.9.4"
   },
   "devDependencies": {
     "@types/express": "^4.17.14",
-    "@types/node": "^18.11.11",
+    "@types/node": "^18.11.12",
     "ts-node": "^10.9.1"
   }
 }
