+++ package.json
@@ -30,9 +30,9 @@
   },
   "devDependencies": {
     "@types/jest": "^29.2.4",
-    "@types/node": "^18.11.11",
-    "@typescript-eslint/eslint-plugin": "^5.45.1",
-    "@typescript-eslint/parser": "^5.45.1",
+    "@types/node": "^18.11.12",
+    "@typescript-eslint/eslint-plugin": "^5.46.0",
+    "@typescript-eslint/parser": "^5.46.0",
     "copyfiles": "^2.4.1",
     "eslint": "^8.29.0",
     "eslint-config-airbnb-base": "^15.0.0",
@@ -42,7 +42,7 @@
     "jest-extended": "^3.2.0",
     "rimraf": "^3.0.2",
     "ts-jest": "^29.0.3",
-    "typescript": "^4.9.3"
+    "typescript": "^4.9.4"
   },
   "dependencies": {
     "statesis": "^0.2.3"
