+++ src/compiler/index.ts
@@ -1,16 +1,15 @@
 import fs from 'fs'
-import { merge, reReference } from 'statesis'
+import { reReference } from 'statesis'
 
-import { parseFile } from './parser'
+import { FileLoaders } from './parser'
+import { getConfig } from './configFileLoader'
 
 import { compileConfig } from './config'
 import { ConvertToSubscribableSchema, NonReserved, Schema } from './types'
 
 export type ConfMonOptions = {
   configDirectory?: string
-  fileLoaders?: {
-    [fileExtension: string]: (data: string, fileName?: string) => Record<string, unknown>
-  }
+  fileLoaders?: FileLoaders
 }
 
 export const compile = <T extends Schema>(
@@ -19,24 +18,14 @@ export const compile = <T extends Schema>(
 ): ConvertToSubscribableSchema<T> => {
   const compiledConfig = compileConfig(schema)
 
-  // TODO: extract to file-load-manager and ensure files processed alphabetically
   const configDirectory = options?.configDirectory ?? './config'
+  const fileLoaders = { ...options?.fileLoaders }
 
-  const getConfig = () => {
-    const configFiles = fs.readdirSync(configDirectory)
-    return configFiles.reduce((acc, file) => {
-      const filePath = `${configDirectory}/${file}`
-      const parsedConfig = parseFile(filePath, { ...options?.fileLoaders })
-
-      return merge(acc, parsedConfig)
-    }, {})
-  }
-
-  let current = getConfig()
+  let current = getConfig(configDirectory, fileLoaders)
   compiledConfig.__applyValue(current)
 
   fs.watch(configDirectory, (_eventType, _filename) => {
-    const newConfig = reReference(current, getConfig())
+    const newConfig = reReference(current, getConfig(configDirectory, fileLoaders))
     current = newConfig
 
     compiledConfig.__applyValue(current)
